requirements.txt
Flask
requests
beautifulsoup4
newspaper3k
extractors
transformers
torch
camel-tools
